+++ src/main/java/aibe1/proj2/mentoss/feature/lecture/controller/LectureController.java
@@ -137,86 +137,4 @@ public class LectureController {
         return ResponseEntity.ok(ApiResponse.ok(responseDto));
     }
 
-    @GetMapping("/{lectureId}/curriculum")
-    @Operation(summary = "강의 커리큘럼 조회", description = "강의의 커리큘럼을 조회합니다.")
-    @ApiResponses(value = {
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                    responseCode = "200",
-                    description = "조회 성공",
-                    content = @Content(schema = @Schema(implementation = ApiResponse.class))
-            ),
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                    responseCode = "404",
-                    description = "강의 없음",
-                    content = @Content(schema = @Schema(implementation = ApiResponse.class))
-            )
-    })
-    public ResponseEntity<ApiResponse<LectureCurriculumResponse>> getLectureCurriculum(@PathVariable Long lectureId) {
-        LectureCurriculumResponse curriculum = lectureService.getLectureCurriculum(lectureId);
-        return ResponseEntity.ok(ApiResponse.ok(curriculum));
-    }
-
-    @GetMapping("/{lectureId}/reviews")
-    @Operation(summary = "강의 리뷰 조회", description = "강의의 리뷰 목록을 조회합니다.")
-    @ApiResponses(value = {
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                    responseCode = "200",
-                    description = "조회 성공",
-                    content = @Content(schema = @Schema(implementation = ApiResponse.class))
-            ),
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                    responseCode = "404",
-                    description = "강의 없음",
-                    content = @Content(schema = @Schema(implementation = ApiResponse.class))
-            )
-    })
-    public ResponseEntity<ApiResponse<LectureReviewsResponse>> getLectureReviews(@PathVariable Long lectureId) {
-        LectureReviewsResponse reviews = lectureService.getLectureReviews(lectureId);
-        return ResponseEntity.ok(ApiResponse.ok(reviews));
-    }
-
-
-
-
-    @PatchMapping("/{lectureId}/status")
-    @Operation(summary = "강의 마감 상태 변경", description = "강의의 마감 상태를 변경합니다.")
-    @ApiResponses(value = {
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                    responseCode = "200",
-                    description = "상태 변경 성공",
-                    content = @Content(schema = @Schema(implementation = ApiResponse.class))
-            ),
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                    responseCode = "404",
-                    description = "강의 없음",
-                    content = @Content(schema = @Schema(implementation = ApiResponse.class))
-            ),
-            @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                    responseCode = "400",
-                    description = "이미 동일한 상태",
-                    content = @Content(schema = @Schema(implementation = ApiResponse.class))
-            )
-    })
-    public ResponseEntity<ApiResponse<Boolean>> updateLectureStatus(
-            @PathVariable Long lectureId,
-            @Parameter(description = "강의 마감 상태", required = true,
-                    examples = {
-                            @ExampleObject(name = "마감", value = "true", summary = "강의를 마감합니다"),
-                            @ExampleObject(name = "오픈", value = "false", summary = "강의를 오픈합니다")
-                    }
-            )
-            @RequestParam boolean isClosed
-    ) {
-        boolean result = lectureService.updateLectureClosed(lectureId, isClosed);
-
-        if (!result) {
-            // 이미 동일한 상태인 경우
-            String message = isClosed ? "이미 마감된 강의입니다." : "이미 오픈된 강의입니다.";
-            return ResponseEntity.badRequest().body(ApiResponse.fail(message));
-        }
-
-        return ResponseEntity.ok(ApiResponse.ok(true));
-    }
-
-
 }
