+++ src/main/java/aibe1/proj2/mentoss/feature/lecture/service/LectureServiceImpl.java
@@ -3,8 +3,9 @@ package aibe1.proj2.mentoss.feature.lecture.service;
 import aibe1.proj2.mentoss.feature.lecture.model.dto.request.LectureCreateRequest;
 import aibe1.proj2.mentoss.feature.lecture.model.dto.request.LectureRegionRequest;
 import aibe1.proj2.mentoss.feature.lecture.model.dto.response.*;
-import aibe1.proj2.mentoss.feature.lecture.model.entity.Lecture;
+import aibe1.proj2.mentoss.global.entity.Lecture;
 import aibe1.proj2.mentoss.feature.lecture.model.mapper.LectureMapper;
+import aibe1.proj2.mentoss.global.entity.Lecture;
 import aibe1.proj2.mentoss.global.exception.EntityNotFoundException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
